Climate entity
Let's welcome the Climate entity :smile:
